+++ Docs/API/html/search/all_0.js
@@ -1,5 +1,6 @@
 var searchData=
 [
   ['allocblocksize',['allocBlockSize',['../struct_jinx_1_1_global_params.html#a94fda3fb37c817db2745cf39f755cead',1,'Jinx::GlobalParams']]],
-  ['allocfn',['allocFn',['../struct_jinx_1_1_global_params.html#aab3f56daccf8ee50fe81df959dfc0fec',1,'Jinx::GlobalParams::allocFn()'],['../_jinx_8h.html#a95bacfc2ada714a6669cda684ed994ac',1,'Jinx::AllocFn()']]]
+  ['allocfn',['allocFn',['../struct_jinx_1_1_global_params.html#aab3f56daccf8ee50fe81df959dfc0fec',1,'Jinx::GlobalParams::allocFn()'],['../_jinx_8h.html#a95bacfc2ada714a6669cda684ed994ac',1,'Jinx::AllocFn()']]],
+  ['allocspareblocks',['allocSpareBlocks',['../struct_jinx_1_1_global_params.html#a82b15a1e12c9acc0a82fbd1a270df003',1,'Jinx::GlobalParams']]]
 ];