+++ Docs/API/html/search/all_a.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['performancestats',['PerformanceStats',['../struct_jinx_1_1_performance_stats.html',1,'Jinx']]]
+  ['performancestats',['PerformanceStats',['../struct_jinx_1_1_performance_stats.html',1,'Jinx']]],
+  ['perftimens',['perfTimeNs',['../struct_jinx_1_1_performance_stats.html#ab17059eff4fac7dc6057bcaa6a85f2fc',1,'Jinx::PerformanceStats']]]
 ];