+++ Docs/API/html/search/variables_7.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['reallocfn',['reallocFn',['../struct_jinx_1_1_global_params.html#ae135b2616b7c2df0f5ec8e12b3f1c9e3',1,'Jinx::GlobalParams']]]
+  ['perftimens',['perfTimeNs',['../struct_jinx_1_1_performance_stats.html#ab17059eff4fac7dc6057bcaa6a85f2fc',1,'Jinx::PerformanceStats']]]
 ];