+++ Source/JxHash.cpp
@@ -2,10 +2,6 @@
 // domain. Appleby disclaims copyright to this source code.
 // https://en.wikipedia.org/wiki/MurmurHash
 
-// Jenkins one at a time hash was written by Bob Jenkins, and has been
-// placed in the public domain.
-// https://en.wikipedia.org/wiki/Jenkins_hash_function
-
 // All code in this source file is compatible with the MIT license.
 
 #include "JxInternal.h"
@@ -57,26 +53,11 @@ uint32_t MurmurHashNeutral2(const void * key, int len, uint32_t seed)
 	return h;
 }
 
-uint32_t jenkins_one_at_a_time_hash(const uint8_t * key, size_t len)
-{
-	uint32_t hash, i;
-	for (hash = i = 0; i < len; ++i)
-	{
-		hash += key[i];
-		hash += (hash << 10);
-		hash ^= (hash >> 6);
-	}
-	hash += (hash << 3);
-	hash ^= (hash >> 11);
-	hash += (hash << 15);
-	return hash;
-}
-
-// Use two 32-bit hashes which are theoretically endian-neutral to reduce odds of collisions
+// Use two endian-neutral 32-bit hashes combined to a 64-bit hash to reduce odds of random collisions
 uint64_t Jinx::GetHash(const uint8_t * data, uint32_t len)
 {
 	uint32_t hash1 = MurmurHashNeutral2(data, len, 0xF835E195);
-	uint32_t hash2 = jenkins_one_at_a_time_hash(data, len);
+	uint32_t hash2 = MurmurHashNeutral2(data, len, 0x5C285D21);
 	uint64_t hash = (static_cast<uint64_t>(hash2) << 32) | static_cast<uint64_t>(hash1);
 	assert(hash != InvalidID);
 	return hash;