+++ Source/JxRuntime.h
@@ -16,6 +16,7 @@ namespace Jinx
 	class Runtime : public IRuntime, public std::enable_shared_from_this<Runtime>
 	{
 	public:
+		Runtime();
 		virtual ~Runtime();
 
 		// IRuntime interface
@@ -50,14 +51,16 @@ namespace Jinx
 
 	private:
 
+		static const size_t NumMutexes = 8;
 		mutable Mutex m_libraryMutex;
 		LibraryMap m_libraryMap;
-		mutable Mutex m_functionMutex;
+		mutable Mutex m_functionMutex[NumMutexes];
 		FunctionMap m_functionMap;
-		mutable Mutex m_propertyMutex;
+		mutable Mutex m_propertyMutex[NumMutexes];
 		PropertyMap m_propertyMap;
 		Mutex m_perfMutex;
 		PerformanceStats m_perfStats;
+		std::chrono::time_point<std::chrono::high_resolution_clock> m_perfStartTime;
 	};